fix(proof): keep the Cobalt ZK guest no_std and gate precompile timing on metrics
evm-std pulled Instant::now() into native precompile dispatch. Cobalt does not need 8130 in the guest, so leave evm-std off until Everest and only start the wall-clock timer when node metrics are enabled.
