+++ etc/systems/tests/eip8130.rs
@@ -2,14 +2,14 @@
 
 #[path = "common/balance.rs"]
 mod balance;
+#[path = "common/zk_dry_run.rs"]
+mod zk_dry_run;
 mod common;
 #[path = "common/zenith.rs"]
 mod zenith;
 
-use std::time::Duration;
-
 use alloy_consensus::Typed2718;
-use alloy_eips::{BlockNumberOrTag, eip2718::Encodable2718};
+use alloy_eips::eip2718::Encodable2718;
 use alloy_network::ReceiptResponse;
 use alloy_primitives::{B256, Bytes, U256};
 use alloy_provider::{Provider, RootProvider};
@@ -19,26 +19,14 @@ use base_common_consensus::{Eip8130Signed, TxEip8130};
 use base_common_network::Base;
 use base_common_rpc_types::BaseTransactionReceipt;
 use base_optimism_rpc::OptimismRollupProviderExt;
-use base_prover_service_client::{ProofRequesterClient, ProverServiceClientConfig};
-use base_prover_service_protocol::{
-    ExecutionStats, GetProofRequest, GetProofResponse, ProofRequest, ProofRequestKind, ProofResult,
-    ProofStatus, ProveBlockRangeRequest, ZkBackend, ZkProofRequest, ZkVm,
-};
 use base_system_tests::{
     ANVIL_ACCOUNT_1, InProcessProverService, InProcessZkHost, SystemTestProviderExt,
     SystemTestStackBuilder,
 };
 use eyre::{Result, WrapErr, ensure};
-use nanoid::nanoid;
-use tokio::time::{sleep, timeout};
-use url::Url;
 
 /// EIP-8130 transaction type byte.
 const EIP8130_TX_TYPE: u8 = 0x79;
-const SAFE_L2_TIMEOUT: Duration = Duration::from_secs(120);
-const SAFE_L2_POLL_INTERVAL: Duration = Duration::from_millis(500);
-const PROOF_TIMEOUT: Duration = Duration::from_secs(15 * 60);
-const PROOF_POLL_INTERVAL: Duration = Duration::from_secs(2);
 
 /// Mines a minimal EOA-path EIP-8130 transaction on the Zenith system-test stack.
 #[tokio::test]
@@ -62,8 +50,10 @@ async fn eip8130_transaction_is_mined() -> Result<()> {
 }
 
 /// Dry-run SP1-executes the block that contains a type `0x79` transaction.
+///
+/// Guest 8130 execution is deferred to Everest (Cobalt ZK stays `no_std`).
 #[tokio::test(flavor = "multi_thread")]
-#[ignore = "SP1 dry-run execute is too slow for merge-queue; run with --release"]
+#[ignore = "8130 guest execution is deferred to Everest; also too slow for merge-queue"]
 async fn eip8130_block_dry_run_proves() -> Result<()> {
     ensure!(
         !cfg!(debug_assertions),
@@ -81,15 +71,20 @@ async fn eip8130_block_dry_run_proves() -> Result<()> {
 
     let rollup_provider =
         RootProvider::<Base>::new_http(system.l2_stack().builder_consensus_rpc_url());
-    wait_for_safe_l2(&rollup_provider, block_number).await?;
+    zk_dry_run::wait_for_safe_l2(&rollup_provider, block_number).await?;
 
     let service = InProcessProverService::start().await?;
     let _host = InProcessZkHost::start(&system, service.url()).await?;
 
     // Pin rollup `head_l1`; local L1 finality lags the batches that made this block safe.
     let l1_head = rollup_provider.optimism_sync_status().await?.head_l1.hash;
-    let stats =
-        prove_block_range_with_dry_run_stats(service.url().clone(), block_number, l1_head).await?;
+    let stats = zk_dry_run::prove_block_range_with_dry_run_stats(
+        service.url().clone(),
+        block_number,
+        l1_head,
+        "eip8130-zk-dry-run",
+    )
+    .await?;
     ensure!(
         stats.total_instruction_cycles > 0,
         "dry-run of an EIP-8130 block must report non-zero instruction cycles"
@@ -142,134 +137,4 @@ async fn send_minimal_eip8130(
         .await
         .wrap_err("EIP-8130 receipt timed out")?;
     Ok((tx_hash, receipt))
-}
-
-async fn wait_for_safe_l2(provider: &RootProvider<Base>, block_number: u64) -> Result<()> {
-    match timeout(SAFE_L2_TIMEOUT, async {
-        loop {
-            let status = provider.optimism_sync_status().await?;
-            if status.safe_l2.number >= block_number {
-                provider.optimism_output_at_block(BlockNumberOrTag::Number(block_number)).await?;
-                return Ok::<_, eyre::Error>(());
-            }
-            sleep(SAFE_L2_POLL_INTERVAL).await;
-        }
-    })
-    .await
-    {
-        Ok(result) => result,
-        Err(_) => {
-            let status = provider.optimism_sync_status().await?;
-            eyre::bail!(
-                "timed out waiting for EIP-8130 block {block_number} to become safe \
-                 (safe_l2={}, unsafe_l2={})",
-                status.safe_l2.number,
-                status.unsafe_l2.number
-            );
-        }
-    }
-}
-
-async fn prove_block_range_with_dry_run_stats(
-    prover_url: Url,
-    block_number: u64,
-    l1_head: B256,
-) -> Result<ExecutionStats> {
-    let start_block_number = block_number
-        .checked_sub(1)
-        .ok_or_else(|| eyre::eyre!("cannot prove genesis block with one-block range"))?;
-    let client_config = ProverServiceClientConfig::new(prover_url.as_str())
-        .with_request_timeout(Duration::from_secs(30));
-    let client = ProofRequesterClient::connect(&client_config)?;
-    let session_id = format!("eip8130-zk-dry-run-{}", nanoid!());
-    let response = client
-        .prove_block_range(ProveBlockRangeRequest {
-            proof: ProofRequest {
-                session_id,
-                request: ProofRequestKind::Compressed(ZkProofRequest {
-                    start_block_number,
-                    number_of_blocks_to_prove: 1,
-                    sequence_window: None,
-                    l1_head: Some(l1_head),
-                    intermediate_root_interval: None,
-                    schedule_l2_block_number: None,
-                    zk_vm: ZkVm::Sp1,
-                    zk_backend: ZkBackend::DryRun,
-                }),
-            },
-            retry_failed: true,
-        })
-        .await?;
-
-    poll_dry_run_stats(&client, response.session_id).await
-}
-
-async fn poll_dry_run_stats(
-    client: &ProofRequesterClient,
-    session_id: String,
-) -> Result<ExecutionStats> {
-    let timeout_session_id = session_id.clone();
-    match timeout(PROOF_TIMEOUT, async {
-        loop {
-            let response =
-                client.get_proof(GetProofRequest { session_id: session_id.clone() }).await?;
-            match response.status {
-                ProofStatus::Succeeded => {
-                    return execution_stats_from_response(&session_id, response);
-                }
-                ProofStatus::Failed => {
-                    return Err(eyre::eyre!(
-                        "proof request failed: {}",
-                        response
-                            .error_message
-                            .unwrap_or_else(|| "missing error message".to_string())
-                    ));
-                }
-                _ => sleep(PROOF_POLL_INTERVAL).await,
-            }
-        }
-    })
-    .await
-    {
-        Ok(result) => result,
-        Err(_) => {
-            let last = client
-                .get_proof(GetProofRequest { session_id: timeout_session_id.clone() })
-                .await
-                .wrap_err_with(|| {
-                    format!(
-                        "timed out waiting for proof request {timeout_session_id}; \
-                         also failed to fetch last status"
-                    )
-                })?;
-            eyre::bail!(
-                "timed out waiting for proof request {timeout_session_id} \
-                 (status={:?}, error={})",
-                last.status,
-                last.error_message.unwrap_or_else(|| "none".to_string())
-            );
-        }
-    }
-}
-
-fn execution_stats_from_response(
-    session_id: &str,
-    response: GetProofResponse,
-) -> Result<ExecutionStats> {
-    match response.result {
-        Some(ProofResult::Compressed(result)) => result.execution_stats.ok_or_else(|| {
-            eyre::eyre!(
-                "dry-run prover response for request {session_id} did not include execution_stats"
-            )
-        }),
-        Some(ProofResult::SnarkPlonk(_)) => Err(eyre::eyre!(
-            "dry-run prover response for request {session_id} returned snark_plonk result"
-        )),
-        Some(ProofResult::Tee(_)) => {
-            Err(eyre::eyre!("dry-run prover response for request {session_id} returned tee result"))
-        }
-        None => Err(eyre::eyre!(
-            "dry-run prover response for request {session_id} did not include a result"
-        )),
-    }
 }
